Remove a "known issue" from the readme.
The issue itself was fixed in Bootstrap 3 somewhere along the way from v3.0.0-RC to v3.2.0.
select2-bootstrap3-css makes use of Bootstrap's own variables to define the height of elements; select2-bootstrap3-css v1.4.0 was the first version not built on Bootstrap 3 RC, but on Bootstrap v3.2.0 – the update of the underlying Bootstrap version resolved the issue.